backend/requirements.txt
# MedixScan Backend Dependencies

# Core Django
Django==4.2.15
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.7
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
cryptography==41.0.4

# API Documentation
drf-yasg==1.21.7

# File Handling & Media
Pillow==10.0.0
django-storages==1.13.2
boto3==1.28.57

# Utilities
python-decouple==3.8
django-extensions==3.2.3
django-filter==23.2
django-import-export==3.2.0

# Async & Background Tasks
celery==5.3.1
redis==4.6.0
django-celery-beat==2.5.0

# Monitoring & Logging
sentry-sdk==1.32.0
django-health-check==3.17.0

# Development
django-debug-toolbar==4.2.0
factory-boy==3.3.0
pytest-django==4.5.2

# Production
gunicorn==21.2.0
whitenoise==6.5.0